#5849 Move test cases to own suite
