+++ core-tests/shared/src/test/scala/zio/CauseSpec.scala
@@ -238,109 +238,6 @@ object CauseSpec extends ZIOBaseSpec {
         val stripped = cause.stripSomeDefects { case _: NumberFormatException => }
         assert(stripped)(isNone)
       }
-    ),
-    suite("captureMultiMethod")(
-      testM("captures a deep embedded failure") {
-        // Given
-        val deepUnderlyingFailure =
-          for {
-            _ <- ZIO.succeed(5)
-            _ <- ZIO.fail("Oh no!").ensuring(ZIO.dieMessage("deep failure"))
-          } yield ()
-
-        val underlyingFailure =
-          for {
-            _ <- ZIO.succeed(15)
-            _ <- deepUnderlyingFailure.ensuring(ZIO.dieMessage("other failure"))
-          } yield ()
-
-        val failure =
-          for {
-            _ <- ZIO.succeed(25)
-            _ <- underlyingFailure
-          } yield ()
-
-        // When
-        val run: ZIO[Any, Throwable, Any] = failure.catchAllCause(cause => ZIO(cause.prettyPrint))
-
-        // Then
-        assertStackTraceMessage(
-          run,
-          includesAll(
-            Seq(
-              "Fiber failed.",
-              "╥",
-              "╠─A checked error was not handled.",
-              "║ Oh no!",
-              "╠─An unchecked error was produced.",
-              "║ java.lang.RuntimeException: deep failure",
-              "╠─An unchecked error was produced.",
-              "║ java.lang.RuntimeException: other failure"
-            )
-          )
-        )
-      },
-      testM("captures the embedded failure") {
-        // Given
-        val underlyingFailure =
-          for {
-            _ <- ZIO.succeed(15)
-            _ <- ZIO.fail("Oh no!").ensuring(ZIO.dieMessage("other failure"))
-          } yield ()
-
-        val failure =
-          for {
-            _ <- ZIO.succeed(25)
-            _ <- underlyingFailure
-          } yield ()
-
-        // When
-        val run: ZIO[Any, Throwable, Any] = failure.catchAllCause(cause => ZIO(cause.prettyPrint))
-
-        // Then
-        assertStackTraceMessage(
-          run,
-          includesAll(
-            Seq(
-              "Fiber failed.",
-              "╥",
-              "╠─A checked error was not handled.",
-              "║ Oh no!",
-              "╠─An unchecked error was produced.",
-              "║ java.lang.RuntimeException: other failure"
-            )
-          )
-        )
-      },
-      testM("captures a single failure if no suppressed cause") {
-        // Given
-        val underlyingFailure =
-          for {
-            _ <- ZIO.succeed(15)
-            _ <- ZIO.fail("Oh no!")
-          } yield ()
-
-        val failure =
-          for {
-            _ <- ZIO.succeed(25)
-            _ <- underlyingFailure
-          } yield ()
-
-        // When
-        val run: ZIO[Any, Throwable, Any] = failure.catchAllCause(cause => ZIO(cause.prettyPrint))
-
-        assertStackTraceMessage(
-          run,
-          (stack: String) =>
-            includesAll(Seq("Fiber failed.\nA checked error was not handled.\nOh no!"))(stack)
-              && excludesAll(
-                Seq(
-                  "An unchecked error was produced",
-                  "other failure"
-                )
-              )(stack)
-        )
-      }
     )
   )
 
@@ -375,19 +272,4 @@ object CauseSpec extends ZIOBaseSpec {
   val throwables: Gen[Random, Throwable] =
     Gen.throwable
 
-  private def includesAll(texts: Seq[String]): String => Boolean = stack => texts.map(stack.contains).forall(r => r)
-
-  private def excludesAll(texts: Seq[String]): String => Boolean = stack => texts.map(stack.contains).forall(!_)
-
-  private def assertStackTraceMessage(
-    failingEffect: ZIO[Any, Throwable, Any],
-    assertion: String => Boolean
-  ): ZIO[Any, Throwable, Assert] =
-    for {
-      stackTrace <- failingEffect
-      result = stackTrace match {
-                 case stackContent: String => assertion(stackContent)
-                 case _                    => false
-               }
-    } yield assertTrue(result)
 }
